Changing types to work with and without pybind1, packaging first working version of population level prediction

+++ src/linear/Utils.h
@@ -30,7 +30,7 @@ class Utils {
 
 class GrayCoder {
     vector<unsigned int> grayCode;
-    vector<uint8_t> reverseGrayCode;
+    vector<unsigned short> reverseGrayCode;
 public:
     explicit GrayCoder(int nBits);
     unsigned int toGrayCode(unsigned int v);
